requirements.txt
streamlit
ollama
pytesseract
Pillow
numpy
httpx>=0.27.0
python-pam
python-dotenv
watchdog
docker
